21_langchain_ecosystem/langgraph/course-notebooks/module-4/studio/requirements.txt
langgraph
langchain-core
langchain-community
langchain-google-genai
tavily-python
wikipedia